# Memorize JSON decoders and encoders