--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -15,7 +15,7 @@ on:
 
 jobs:
   lint:
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-latest
     timeout-minutes: 30
     strategy:
       fail-fast: false
@@ -25,14 +25,16 @@ jobs:
         with:
           fetch-depth: 0
       - name: Setup Scala and Java
-        uses: olafurpg/setup-scala@v13
+        uses: coursier/setup-action@v1
+        with:
+          apps: sbt
       - name: Cache scala dependencies
         uses: coursier/cache-action@v6
       - name: Lint code
         run: sbt fmtCheck fixCheck
 
   build:
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-latest
     timeout-minutes: 40
     strategy:
       fail-fast: false
@@ -47,6 +49,7 @@ jobs:
     - uses: coursier/setup-action@v1
       with:
         jvm: ${{ matrix.java }}
+        apps: sbt
     - name: Cache scala dependencies
       uses: coursier/cache-action@v6
     - name: Install Bohem GC
@@ -56,13 +59,13 @@ jobs:
       run: sbt ++${{ matrix.scala }}! test${{ matrix.platform }}
 
   ci:
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-latest
     needs: [build,lint]
     steps:
       - run: echo "All checks passed"
 
   publish:
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-latest
     timeout-minutes: 45
     needs: [build,lint]
     if: github.event_name != 'pull_request'
@@ -70,7 +73,9 @@ jobs:
       - uses: actions/checkout@v3.0.0
         with:
           fetch-depth: 0
-      - uses: olafurpg/setup-scala@v13
+      - uses: coursier/setup-action@v1
+        with:
+          apps: sbt
       - run: sbt ci-release
         env:
           PGP_PASSPHRASE: ${{ secrets.PGP_PASSPHRASE }}