--- a/.github/workflows/site.yml
+++ b/.github/workflows/site.yml
@@ -21,12 +21,9 @@ jobs:
       uses: actions/checkout@v3.3.0
       with:
         fetch-depth: '0'
-    - name: Setup Scala
-      uses: actions/setup-java@v3.9.0
+    - uses: coursier/setup-action@v1
       with:
-        distribution: temurin
-        java-version: 17
-        check-latest: true
+        apps: sbt
     - name: Check artifacts build process
       run: sbt  +publishLocal
     - name: Check website build process
@@ -40,12 +37,9 @@ jobs:
       uses: actions/checkout@v3.3.0
       with:
         fetch-depth: '0'
-    - name: Setup Scala
-      uses: actions/setup-java@v3.9.0
+    - uses: coursier/setup-action@v1
       with:
-        distribution: temurin
-        java-version: 17
-        check-latest: true
+        apps: sbt
     - name: Setup NodeJs
       uses: actions/setup-node@v3
       with:
@@ -65,12 +59,9 @@ jobs:
       with:
         ref: ${{ github.head_ref }}
         fetch-depth: '0'
-    - name: Setup Scala
-      uses: actions/setup-java@v3.9.0
+    - uses: coursier/setup-action@v1
       with:
-        distribution: temurin
-        java-version: 17
-        check-latest: true
+        apps: sbt
     - name: Generate Readme
       run: sbt  docs/generateReadme
     - name: Commit Changes